third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/clear-expected.txt

This is a testharness.js-based test.
[FAIL] Can clear a CSS rule containing properties
  assert_equals: Custom properties should be cleared expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
